Cognitive Test to Aid in the Detection of Early-Onset Alzheimer's Disease
This program works in tandem with the TimeStats Google Chrome
application to administer a personalized, repeatable cognitive test
that can be used to track cognitive ability in the long term. The test
is based off of a user's internet history and can therefore be repeated
daily. The program is designed to aid in the detection of early-onset
Alzheimer's Disease.